+++ program/sensors_math.cpp
@@ -8,11 +8,11 @@
 /////////////////////////////
 // GENERIC MATH FUNCTIONS
 
-float faverage(float* array, int dim)
+double faverage(double* array, int dim)
 {
     return 3;
 }
-float fvariance(float* array, int dim)
+double fvariance(double* array, int dim)
 {
     return 3;
 }
